requirements.txt
fastapi==0.92.0
openai==0.26.5
pydantic==1.10.5
python-dotenv==0.21.1
uvicorn==0.12.3
aiofiles==0.6.0
